esm fix coverage 1.0.7 sync wallet-lib
