+++ packages/lorena-sdk/index.js
@@ -1,0 +1,3 @@
+// Set options as a parameter, environment variable, or rc file.
+require = require('esm')(module/*, options */)
+module.exports = require('./src/index.js')
